src/main/java/backend/model/question/Question.java
package backend.model.question;

import com.google.gson.annotations.SerializedName;

import javax.persistence.*;

@Entity
@Inheritance(strategy=InheritanceType.SINGLE_TABLE)
public abstract class Question {

    private Long id;
    @SerializedName("question")
    private String questionText;

    @Id
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getQuestionText() {
        return questionText;
    }

    public void setQuestionText(String questionText) {
        this.questionText = questionText;
    }
}
